moved utiliity functions to separate file
