+++ studio/components/interfaces/Settings/Logs/index.ts
@@ -1,5 +1,6 @@
 export * from './Logs.types'
 export * from './Logs.constants'
+export * from './Logs.utils'
 export { default as LogsSavedQueriesItem } from './Logs.SavedQueriesItem'
 export { default as LogsQueryPanel } from './LogsQueryPanel'
 export { default as LogTable } from './LogTable'
